organzing chat componenets
