+++ frontend/src/components/asidebar/asidebar.tsx
@@ -46,7 +46,7 @@ const AsideBar = () => {
           </li>
           <li>
             <Link
-              href="#"
+              href="/chat"
               className="flex items-center p-2 text-gray-900 rounded-lg dark:text-white hover:bg-gray-100 dark:hover:bg-gray-700 group"
             >
               <BsChatRightDots />
